requirements.txt
Click==7.0
cycler==0.10.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
mysql-connector-python==8.0.18
numpy==1.17.4
pandas==0.25.3
protobuf==3.11.1
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
Werkzeug==0.16.0
